+++ packages/web-sdk/src/instrumentations/userActions/httpRequestMonitor.ts
@@ -1,126 +1,125 @@
-import { Observable } from '@grafana/faro-core';
+import { genShortID, Observable } from '@grafana/faro-core';
 
 import { getUrlFromResource, isUrlIgnored } from '../../utils/url';
 
 import { MESSAGE_TYPE_HTTP_REQUEST_END, MESSAGE_TYPE_HTTP_REQUEST_START } from './const';
-import type { HttpRequestEndMessage, HttpRequestStartMessage } from './types';
+import type { HttpRequestEndMessage, HttpRequestMessagePayload, HttpRequestStartMessage } from './types';
+
+type RequestProps = HttpRequestMessagePayload;
+
+const apiTypeFetch = 'fetch';
+const apiTypeXhr = 'xhr';
 
 /**
  * Monitors if any http requests are in progress.
  */
-export function monitorHttpRequests(): {
-  observable: Observable<HttpRequestStartMessage | HttpRequestEndMessage>;
-  resetCounters: () => void;
-} {
+export function monitorHttpRequests(): Observable<HttpRequestStartMessage | HttpRequestEndMessage> {
   const observable = new Observable<HttpRequestStartMessage | HttpRequestEndMessage>();
 
-  let pendingXhrRequests = 0;
-  let pendingFetchRequests = 0;
-
-  function emitStartMessage() {
-    observable.notify({ type: MESSAGE_TYPE_HTTP_REQUEST_START, pending: pendingXhrRequests + pendingFetchRequests });
+  function emitStartMessage(requestProps: RequestProps) {
+    observable.notify({
+      type: MESSAGE_TYPE_HTTP_REQUEST_START,
+      request: requestProps,
+    });
   }
 
-  function emitEndMessage() {
-    observable.notify({ type: MESSAGE_TYPE_HTTP_REQUEST_END, pending: pendingXhrRequests + pendingFetchRequests });
+  function emitEndMessage(requestProps: RequestProps) {
+    observable.notify({
+      type: MESSAGE_TYPE_HTTP_REQUEST_END,
+      request: requestProps,
+    });
   }
 
   monitorFetch({
-    onRequestStart: () => {
-      pendingFetchRequests++;
-      emitStartMessage();
-    },
-    onRequestEnd: () => {
-      pendingFetchRequests--;
-      emitEndMessage();
-    },
+    onRequestStart: emitStartMessage,
+    onRequestEnd: emitEndMessage,
   });
 
   monitorXhr({
-    onRequestStart: () => {
-      pendingXhrRequests++;
-      emitStartMessage();
-    },
-    onRequestEnd: () => {
-      pendingXhrRequests--;
-      emitEndMessage();
-    },
+    onRequestStart: emitStartMessage,
+    onRequestEnd: emitEndMessage,
   });
 
-  function resetCounters() {
-    pendingXhrRequests = 0;
-    pendingFetchRequests = 0;
-  }
-
-  return { observable, resetCounters };
+  return observable;
 }
 
-function monitorXhr({ onRequestStart, onRequestEnd }: { onRequestStart: () => void; onRequestEnd: () => void }) {
+function monitorXhr({
+  onRequestStart,
+  onRequestEnd,
+}: {
+  onRequestStart: (props: RequestProps) => void;
+  onRequestEnd: (props: RequestProps) => void;
+}) {
   const originalOpen = XMLHttpRequest.prototype.open;
-  const originalSend = XMLHttpRequest.prototype.send;
 
   XMLHttpRequest.prototype.open = function () {
     const url = arguments[1];
     const isIgnoredUrl = isUrlIgnored(url);
+    const method = arguments[0];
 
+    const requestId = genShortID();
+
+    // request has started to load data.
+    this.addEventListener('loadstart', function () {
+      if (!isIgnoredUrl) {
+        onRequestStart({ url, method, requestId, apiType: apiTypeXhr });
+      }
+    });
+
+    // transaction completes successfully.
     this.addEventListener('load', function () {
       if (!isIgnoredUrl) {
-        onRequestEnd();
+        onRequestEnd({ url, method, requestId, apiType: apiTypeXhr });
       }
     });
 
     this.addEventListener('error', function () {
       if (!isIgnoredUrl) {
-        onRequestEnd();
+        onRequestEnd({ url, method, requestId, apiType: apiTypeXhr });
       }
     });
 
     this.addEventListener('abort', function () {
       if (!isIgnoredUrl) {
-        onRequestEnd();
+        onRequestEnd({ url, method, requestId, apiType: apiTypeXhr });
       }
     });
 
     originalOpen.apply(this, arguments as any);
   };
-
-  XMLHttpRequest.prototype.send = function () {
-    const url = getUrlFromResource(arguments[0]);
-
-    console.log('url', url);
-
-    const isIgnoredUrl = isUrlIgnored(url);
-
-    if (!isIgnoredUrl) {
-      onRequestStart();
-    }
-
-    originalSend.apply(this, arguments as any);
-  };
 }
 
-function monitorFetch({ onRequestEnd, onRequestStart }: { onRequestStart: () => void; onRequestEnd: () => void }) {
+function monitorFetch({
+  onRequestEnd,
+  onRequestStart,
+}: {
+  onRequestStart: (props: RequestProps) => void;
+  onRequestEnd: (props: RequestProps) => void;
+}) {
   const originalFetch = window.fetch;
 
   window.fetch = function () {
-    const url = getUrlFromResource(arguments[0]);
+    const url = getUrlFromResource(arguments[0]) ?? '';
     const isIgnoredUrl = isUrlIgnored(url);
+    const method = (arguments[1] ?? {}).method;
+
+    const requestId = genShortID();
 
     if (!isIgnoredUrl) {
-      onRequestStart();
+      onRequestStart({ url, method, requestId, apiType: apiTypeFetch });
     }
 
     return originalFetch
       .apply(this, arguments as any)
       .then((response) => {
         if (!isIgnoredUrl) {
-          onRequestEnd();
+          onRequestEnd({ url, method, requestId, apiType: apiTypeFetch });
         }
         return response;
       })
       .catch((error) => {
         if (!isIgnoredUrl) {
-          onRequestEnd();
+          onRequestEnd({ url, method, requestId, apiType: apiTypeFetch });
         }
         throw error;
       });